rename module logging to print
